refactor: reorganize UDP server mod
