+++ tests/servers/udp/mod.rs
@@ -1,7 +1,7 @@
-use torrust_tracker::servers::udp::server;
+use torrust_tracker::servers::udp::server::states::Running;
 
 pub mod asserts;
 pub mod contract;
 pub mod environment;
 
-pub type Started = environment::Environment<server::Running>;
+pub type Started = environment::Environment<Running>;
